(PPE) Change data type of model fields

+++ Module.Persist.TPM/Model/TPM/PlanPostPromoEffect.cs
@@ -15,8 +15,8 @@ namespace Module.Persist.TPM.Model.TPM
 
         [StringLength(255)]
         public string Size { get; set; }
-        public decimal PlanPostPromoEffectW1 { get; set; }
-        public decimal PlanPostPromoEffectW2 { get; set; }
+        public double PlanPostPromoEffectW1 { get; set; }
+        public double PlanPostPromoEffectW2 { get; set; }
         
         public Guid DiscountRangeId { get; set; }
         public DiscountRange DiscountRange { get; set; }
